requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pandas
pillow
requests
python-multipart
